+++ packages/nimble-tokens/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-tokens
 
-This log was last generated on Tue, 30 Jul 2024 14:34:23 GMT and should not be manually modified.
+This log was last generated on Thu, 01 Aug 2024 08:51:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 7.2.0
+
+Thu, 01 Aug 2024 08:51:10 GMT
+
+### Minor changes
+
+- Add color tokens for mobiscroll components ([ni/nimble@2488638](https://github.com/ni/nimble/commit/2488638308b5e511f247dc1d113f1b2b6d74c2d4))
 
 ## 7.1.0